casting  sql count from Long to Integer

+++ eBudgeting/src/main/java/biz/thaicom/eBudgeting/repositories/ActivityTargetReportRepository.java
@@ -128,7 +128,7 @@ public interface ActivityTargetReportRepository extends
 			+ "FROM ActivityTargetReport report "
 			+ "WHERE report.target.id=?1 and report.owner.parent.id = ?2 "
 			+ "")
-	public Integer countAllByTarget_IdAndOwner_Parent_id(Long activityTargetId, Long parentOrgId);
+	public Long countAllByTarget_IdAndOwner_Parent_id(Long activityTargetId, Long parentOrgId);
 	
 	public List<ActivityTargetReport> findAllByTarget_IdAndOwner_Parent_id(
 			Long activityTargetId, Long parentOrgId);
